fix: fix "n_sia" logic of `add_glycan_description`
This bug comes from glyrepr v0.4.2, where `glyrepr::count_mono()` returns NA when the glycan has "generic" type and the mono to test has "concrete" type. "Sia" is regarded as a "concrete" one so `count_mono(glycan, "Sia")` always returns NA in our situation.

This bug should be fix in the `glyrepr` packge, in that "Sia" should be taken special care.

For now, we just assumes "Sia" is not common and we simply add up "NeuAc" and "NeuGc" to get "Sia".